Distributed Machine Learning (Supervised)
0.1 Dataset
You have to use two datasets for this task. Read the description of the dataset. You have to use them
for learning a regression model.
• Dynamic Features of VirusShare Executables Data Set https://archive.ics.uci.edu/ml/datasets/
Dynamic+Features+of+VirusShare+Executables
• KDD Cup 1998 Data Data Set https://archive.ics.uci.edu/ml/datasets/KDD+Cup+1998+
Data
Please divide your data into 70% train and 30% test. You will train your model on train part of the data
and will test it on test part of the data. Use RMSE train and test score for measuring the performance
of your model.