Image Classification with TensorFlow
Image classification is a task of assigning an input image I ∈ R
H×W×C a class l out of a fixed set of
given classes L ∈ {1 . . . L}, where H is the height, W is the weight and C is the number of channels, L
are the number of classes. For example an input image [32 × 32 × 3] has height 32, weight 32 and three
channels (a color image has RGB values). In this task we will look at CIFAR-10 dataset available at
https://www.cs.utoronto.ca/~kriz/cifar.html. CIFAR-10 dataset consists of 60000 color images
(each image is a 32 × 32 × 3. The data is already divided into five batches and test batch. You can
download the data and load it into python dict object as explained on the said URL i.e