Pandas and Numpy
• Word Count Program: In this task your are required to use the provided text
file and write a program that will count the number of occurances of unique words.
The program should ignore words like {’the’, ’a’,’an’,’be’}
Finally you are required to generate the histogram of the top 10 most occuring
words. Click here to download Text File
• Matrix Multiplication: Using numpy you are required to use numpy for operation on matrices. Create a matrix A of dimensions n × m, where n = 100 and
m = 20. Initialize Matrix A. Create a vector v of dimension m × 1. Initialize the
matrix with a random values and vector with normal distribution using µ = 2 and
σ = 0.01. Perform following operation on them
– Iterative multiply (element-wise) each row of matrix A with vector v and sum
the result of each iteration in another vector c
– Find mean and standard deviation of the new vector c
– Plot histogram of vector c using 5 bins